Style apps menu
