+++ src/front/ui/components/modals/style.scss
@@ -69,11 +69,13 @@
     right:0;
     bottom:0;
     background: var(--fl-asst-overlay-background-color);
+    overflow: hidden;
 
     .fl-asst-screen-menu-contents {
         background: var(--fl-background-color);
         min-width: 440px;
         margin: 0 auto;
         flex: 0 0 auto;
+        box-shadow: 0 12px 24px 0 rgba(0,0,0,0.17);
     }
 }
