+++ src/front/ui/system/apps-menu/style.scss
@@ -1,0 +1,25 @@
+.fl-asst-app-list {
+    display: grid;
+    padding: 18px 0;
+}
+.fl-asst-app-list-item {
+    display: flex;
+    flex-direction: row;
+    padding: 12px 16px;
+    padding-left: 30px;
+    cursor: pointer;
+
+    &:hover {
+        background: var(--fl-utility-background-color);
+    }
+}
+.fl-asst-manage-apps-title {
+    margin:0;
+    padding: 30px 30px 0;
+}
+.fl-asst-app-list-item-title {
+    font-size: 16px;
+    font-weight: 600;
+    margin-left: 16px;
+    line-height: 1.4;
+}
