--- src/front/ui/system/style.scss
@@ -1,13 +1,0 @@
-.fl-asst-app-list {
-    display: grid;
-    padding: 20px 0;
-}
-.fl-asst-app-list-item {
-    display: flex;
-    flex-direction: row;
-    padding: 10px 16px;
-    cursor: pointer;
-}
-.fl-asst-manage-apps-title {
-    padding: 30px 30px 0;
-}
